+++ routes/builderRoutes/archive.js
@@ -37,22 +37,58 @@ module.exports = function (req, res) {
                     var obj = new ArchiveOther(item.toObject())
                     obj.isNew = true
                     obj.type = type.slice(1)
-                    console.log('new archive obj', obj)
-                    obj.save(function (err, newItem) {
-                        if (err) {
-                            console.log('Archive not added...', err)
-                            res.status(400).send('Archive not added...')
-                        } else {
-                            mediaList[type].deleteOne({ _id: id }, function (err) {
-                                if (err) {
-                                    console.log('Media not deleted...', err)
-                                    res.status(400).send('Media not deleted...')
-                                } else {
-                                    console.log('archived!')
-                                    res.send(newItem)
+                    firebase.db.ref('o/' + type.slice(1) + 's/' + id).once('value', (content) => {
+                        console.log('fb content', content.val())
+                        console.log('new archive obj', obj)
+                        obj.save(function (err, newItem) {
+                            if (err) {
+                                console.log('Archive not added...', err)
+                                res.status(400).send('Archive not added...')
+                            } else {
+                                console.log('before structure', newItem)
+                                newItem.content = {}
+                                newItem.content.structure = content.val().structure
+                                if (content.val().modules) {
+                                    console.log('run modules')
+                                    newItem.content.modules = Object.entries(content.val().modules).map((entry) => {
+                                        entry[1].key = entry[0]
+                                        return entry[1]
+                                    })
                                 }
-                            })
-                        }
+                                if (content.val().sections) {
+                                    console.log('run sections')
+                                    newItem.content.sections = Object.entries(content.val().sections).map((entry) => {
+                                        entry[1].key = entry[0]
+                                        return entry[1]
+                                    })
+                                }
+                                if (content.val().sectionModules) {
+                                    console.log('run sectionModules')
+                                    newItem.content.sectionModules = Object.entries(content.val().sectionModules).map((entry) => {
+                                        entry[1].key = entry[0]
+                                        return entry[1]
+                                    })
+                                }
+                                newItem.save(function (err, finalItem) {
+                                    if (err) {
+                                        console.log('Archive not added...', err)
+                                        res.status(400).send('Archive not added...')
+                                    } else {
+                                        mediaList[type].deleteOne({ _id: id }, function (err) {
+                                            if (err) {
+                                                console.log('Media not deleted...', err)
+                                                res.status(400).send('Media not deleted...')
+                                            } else {
+                                                firebase.db.ref('o/' + type.slice(1) + 's/' + id).remove()
+                                                console.log('archived!')
+                                                res.send(finalItem)
+                                            }
+                                        })
+                                    }
+                                })
+                                
+                            }
+                        })
                     })
                 })
             } else {