+++ routes/routes.js
@@ -24,6 +24,7 @@ const builderbible = require('./builderRoutes/bible.js')
 const builderresources = require('./builderRoutes/resources.js')
 const builderresearch = require('./builderRoutes/research.js')
 const builderuser = require('./builderRoutes/user.js')
+const builderpoll = require('./builderRoutes/poll.js')
 
 app.use(function(req, res, next) {
     res.header('Access-Control-Allow-Origin', '*')
@@ -65,5 +66,6 @@ app.post('/builder/search', buildersearch)
 app.post('/builder/bible', builderbible)
 app.post('/builder/resources', builderresources)
 app.post('/builder/research', builderresearch)
+app.post('/builder/poll', builderpoll)
 
 module.exports = app